split harn-vm schema module
